updated module implementation

+++ src/main/java/technology/positivehome/ihome/domain/runtime/sensor/Dds238PowerMeterData.java
@@ -1,3 +1,3 @@
 package technology.positivehome.ihome.domain.runtime.sensor;
-public record Dds238PowerMeterData(Integer voltage) {
+public record Dds238PowerMeterData(Double voltage) {
 }
